+++ src/modules/backup/backup.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { PostgresBackupGateway } from './postgres-backup.gateway';
 import { BackupService } from './backup.service';
 
 @Module({
-  providers: [BackupService],
+  providers: [PostgresBackupGateway, BackupService],
   exports: [BackupService],
 })
 export class BackupModule {}